[Manju]:refactor login
