+++ src/App.tsx
@@ -5,8 +5,8 @@ import {
   BrowserRouter as Router,
   Routes,
 } from "react-router-dom";
-import Login from "./Login";
-import Signup from "./Signup";
+import Login from "./components/Login";
+import Signup from "./components/Signup";
 
 function App() {
   return (
